report fix

+++ finance/egov/egov-commons/src/main/java/org/egov/commons/Fund.java
@@ -72,13 +72,13 @@ import org.hibernate.validator.constraints.SafeHtml;
 @Table(name = "fund")
 @SequenceGenerator(name = Fund.SEQ, sequenceName = Fund.SEQ, allocationSize = 1)
 @Unique(fields = {"code", "name"}, enableDfltMsg = true)
-public class Fund extends AbstractPersistable<Integer> {
+public class Fund extends AbstractPersistable<Long> {
 
     public static final String SEQ = "SEQ_Fund";
     private static final long serialVersionUID = 7977534010758407945L;
     @Id
     @GeneratedValue(generator = Fund.SEQ, strategy = GenerationType.SEQUENCE)
-    private Integer id;
+    private Long id;
 
     @Length(max = 50, min = 2)
     @SafeHtml
@@ -111,11 +111,11 @@ public class Fund extends AbstractPersistable<Integer> {
 
     private Date lastModifiedDate;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
